Merged PR 62997: Fix config path for ddg_automation_cleaner cron
## What's being changed

The cleaner cron config path.

## Why it's being changed

In [this commit](https://github.com/dotmailer/dotmailer-magento2-extension/commit/1b9301b51be2a90247275232b4f1563c9fe94641) we changed the config paths for the cleaner settings, but didn't amend the crontab. The impact is that the cleaner cron does not run in 4.28.0 and 4.29.0.

## How to review / test this change

- Turn on cron
- Set cleaner cron to run once per day (Developer > Cleaner)
- Clear cache
- Observe a ddg_automation_cleaner is scheduled after max 24hrs

Related work items: #290045