Merged PR 63548: Move Logger from inheritance to composition
## What's being changed
We have refactored the `Dotdigitalgroup\Email\Logger\Logger` class from a Inheritance design to a composition design.

We have added a new `Dotdigitalgroup\Email\Api\Logger\LoggerInterface` and preference respectfully to indicate that the logger will not change its structure  in the foreseeable future.

## Why it's being changed
In Magento 2.4.8, monolog/monolog 3+ is bundled, meaning \Monolog\Logger is now marked as @final. We could simply skip the PHPStan error and carry on, but I felt it might be preferable to follow the Adobe docs  in this case.

## How to review / test this change
- Run setup:upgrade
- Run any sync, confirm logs are output as expected

Related work items: #292288