Merged PR 63783: Make catalog sync type and visibility filters store-scoped
## What's being changed

The settings for what products types and product visibilities have been changed to be store-scoped.

## Why it's being changed

Catalogs (unless we're syncing a single default-level catalog) are synced by store. So it doesn't make much sense to inject StoreManager to get a parent website id, when we can just allow the setting to inherit - or, allow merchants to sync different product types and visibilities into different catalogs, if they so wish.

## How to review / test this change

- Reset your catalog
- Change the settings to sync e.g. all types for one store, simple products only for another store
- Sync catalogs at store level
- Reset again and sync at default level
- See no errors

Related work items: #293531