Merged PR 64160: Fix PHPStan error in Connector\Product
## What's being changed

Addresses this error that surfaces in PHPStan in PHP 8.4:
```
Cannot unset property Dotdigitalgroup\Email\Model\Connector\Product::$attributes because it might have hooks in a subclass.
```

## How to review / test this change

- Export a product for catalog sync
- If the product has no attributes, we must not see `attributes` in the JSON

## Notes

If attributes is passed as null instead of as an array, this will lead to an error with insight data. It must be removed from the export.

Related work items: #295274