Merged PR 64451: Abandoned Cart guest enrolment - fix contact creation
## What's being changed
Contact creation moved before automation is published to ensure contact is created before automation is processed

## Why it's being changed
Abandoned cart enrolments for guests can fail (with the message 'Could not find matching contact in email_contact.') if the enrolment happens before the contact is created locally off the back of `getOrCreateContactByEmail`. This only started happening when we moved automations to the queue.

## How to review / test this change
- Enable AC program enrolment
- Enable Email Capture
- Allow AC for non-subscribers
- As a guest, add a product to the cart
- Enter your email address at checkout and abandon
- Validate contact is enrolled in `email_automation` table

Related work items: #293333