Merged PR 64416: Set upper bound for magento/framework compatibility to 2.4.8
## What's being changed

We've tightened the compatibility range for Magento support to the 2.4.8 line. This will include any patch releases for 2.4.8, but will not include 2.4.9 until we change this again.

## Why it's being changed

Adobe may release significant, breaking changes in a minor version release that they do not include in a preceding beta. Setting an upper bound in our dependencies allows us time to test a release and confirm compatibility, instead of scrambling to fix something that is already broken in active installations.

## How to review / test this change

- Check the updated version constraint against the [version used in 2.4.8](https://github.com/magento/magento2/blob/2.4.8/composer.json#L363)
- Confirm our revised constraint will allow patch releases fx 103.0.8-p1

Related work items: #296484