Merged PR 64677: Delay new subscriber automation publish until after subscriber is created
## What's being changed

This PR gives the SubscriptionConsumer responsibility for publishing a New Subscriber automation.

## Why it's being changed

We had reports of new subscriber automation enrolments occasionally failing with this odd message from the API:
```
"error_message":"The email must be unique. ERROR_NON_UNIQUE_EMAIL"
```
The problem turned out to be a race condition where both the SubscriptionConsumer and the AutomationConsumer were trying to create or get the same contact at roughly the same time.

The sequence was:
- Contact subscribes in Magento
- Connector pushes them to subscriber list via V3 PATCH request
- Eligible for subscriber automation (which in Magento requires a pre-sync step) so
- Connector pushes them to ‘All Contacts’ via V2 POST to https://r2-api.dotmailer.com/v2/contacts/with-consent-and-preferences
- fails

V2 couldn't find the contact yet, so it tried to create it too, but couldn't because the creation was to some degree already in play.

The resolution is to wait until after the contact is subscribed in the SubscriptionConsumer before queuing the automation.

## How to review / test this change

- Run setup:upgrade (this is required to properly set the new SubscriptionData model structure for some reason)
- Configure subscriber automation
- Subscribe an email address
- See a row created in email_automation
- See a row created in email_contact
- See a message queued for `ddg.newsletter.subscription` with an automation_id
- See a message queued for `ddg.sync.automation`
- See a subscriber created in Dotdigital and enrolled to a program

## Notes
I believe this is the only case where we have sequential queue consumers running that need to create a contact, then get the same contact. Other similar contexts I have checked are:
- OrderPlaceAfter - cartPhaseUpdate + enrol (ok because this is an insight data upsert for an existing contact)
- OrderSaveAfter - subscribe + enrol (this is a reset/resync for an existing subscriber so would not involve creating a contact)

Related work items: #297207