Merged PR 65003: Only use default for null attribute values
## What's being changed

When using default values for attributes, we will only use the default if the found value is `null` (not just false-y).

## Why it's being changed

There was a regression from [this change](https://github.com/dotmailer/dotmailer-magento2-extension/commit/9fccc9d5a3d3b51431a353ef903fe4b4f05a102d) where Yes/No attributes were syncing with a '1' value incorrectly, when their actual value was '0'. '0' is of course false-y so the code thought this needed to use the default.

## How to review / test this change

- Create a Yes/No customer attribute with the Default Value of 'Yes'
- Map this attribute for sync via Data Mapping > Custom Attributes
- Pick three customers:
- Customer 1 is not changed/saved (i.e. they use the default state)
- Customer 2 is saved with 'Yes'
- Customer 3 is saved with 'No'
- Ensure customers 1 and 2 are reset for sync
- Run customer sync

The expected result is:
```
Customer 1 is exported with '1'
Customer 2 is exported with '1'
Customer 3 is exported with '0'
```
- Retry the above after setting the Default Value of the attribute to 'No'

The expected result is:
```
Customer 1 is exported with '0'
Customer 2 is exported with '1'
Customer 3 is exported with '0'
```

Related work items: #299183