Merged PR 59380: MFTF test fixes
## What's being changed

This PR is the result of a long-term campaign to wrestle the world's most brittle and unpredictable test framework into a useable CI test suite.

## Why it's being changed

It's taken a long time, mainly because for a while, we couldn't get the actual pipeline running with newer versions of Magento and magento/magento-cloud-docker. It was also very difficult, because until recently we had no access to screenshots from failed tests. Once we published these as artifacts it became significantly easier.

Some highlights from the test fixes:
- replace window.onLoad with domReady in setupIntegration.js. This fixed a problem where the account page loaded but the button click did nothing.
- all tests where a product was created needed to a) add the product to the default source and b) run a reindex to ensure the products could be added to the orders created in the admin.
- allow the cron step of the SetupIntegration test to show a success or a failure depending on the test environment (when Magento runs it, cron is active, in our CI pipeline it isn't)
- added action groups to set and clear the dynamic content passcode

## How to review / test this change

- See a green build for MFTF

Related work items: #272733