Merged PR 66438: Add catalogrule_product and catalog_product_price to list of deferring indexers
## What's being changed

The `CatalogSyncDeferralHandler` now also checks the `catalogrule_product` indexer and the `catalog_product_price` indexer (as well as the `catalogrule_rule` indexer) to see if they have pending updates, when deciding whether to defer catalog sync.

## Why it's being changed

These additional indexers also recalculate data in the `catalog_product_index_price` table, which we load index prices from directly. Even if there are no catalog price rules set up, it is expected that for example, updating the product price, or adding a special price on a product, will also be reflected in the index prices, if these are being synced.

## How to review / test this change

- Turn on syncing index prices
- Put a breakpoint in the CatalogSyncDeferralHandler
- Make a change to a product's price
- Run catalog sync immediately
- See catalog sync deferred
- Repeat for setting a special price
- Test any other direct price amendment on a product

## Notes

Use bin/magento indexer:status to see the impact of price changes on product-related indexers.

Related work items: #304765