Merged PR 65515: Customer batch syncing: continue if batch is empty
## What's being changed

Customer sync batching logic.

## Why it's being changed

In customer sync, it is possible that we try to get a customer collection for a batch of e.g. 500 ids, but don't find any syncable records. We have seen this happen in cases where merchants have seeded large amounts of sample or dummy data for customers into `email_contact`. In such cases, if we break on an empty batch, we will never sync the full database of contacts.

As long as keep updating the offset, we can instead `continue` if a batch is empty, and proceed to try the next set of customer ids.

## How to review / test this change

- Set contact sync batch size to 10 in Developer > Import Settings
- Reset your contacts
- Make a change to the code in Model/Sync/Customer/CustomerDataManager L70:
```
if (max($customerIds) > 20) {
    // run the code as it is normally
    $customerCollection->addAttributeToFilter('entity_id', ['in' => $customerIds]);
} else {
    // create an empty collection for testing purposes
    $customerCollection->addFieldToFilter('entity_id', ['in' => [1000000, 1000001, 1000002]]);
}
```
- Run customer sync
- See in the logs e.g.
```
dotdigital.DEBUG: Could not batch customers in website 1. Tried exporting 10 ids from 1 to 18. Continuing.
```
- See the remainder of your customers successfully batched in `email_importer`
- See all customer rows marked as email_imported apart from the skipped rows.

## Notes

There is no 'processed' state for contacts in `email_contact` as distinct from 'imported', like there is in other tables. Customers are either imported or they aren't, and if they aren't, they are candidates for the next sync. So if we can't create a collection from dummy rows, we will keep trying on every sync, and outputting the same log message. I'm not inclined to write further code to work around this, since ultimately it's a sign that the data is bad.

Related work items: #301488