Merged PR 67434: Surface original exception message in coupon request processor
## What's being changed

DotdigitalCouponRequestProcessor (exception handling).

## Why it's being changed

In `generateNewCoupon` we are catching the `LocalizedException` and throwing our own `ErrorException`. Unfortunately we therefore don't get to see the reason for the `LocalizedException`.

## How to review / test this change

- Set up a cart price rule
- Set coupon to 'Auto'
- Use the EDC URL to generate a coupon against an email address
- See a relevant error message in the logs
- Switch the rule back to 'Specific Coupon' with 'Auto-generation' checked
- Re-run the EDC, see a coupon

## Notes

The fact 'Auto' is not allowed is clearly confusing, but allowing our request processor to handle this as well is a story for another day.

Related work items: #309851