Merged PR 67210: Optimise InsertEmailContactTableSubscribers data migration
## What's being changed

The `joinLeft` in the select is replaced with a `NOT EXISTS`. I've also removed the sort on `subscriber_id`, which is not necessary on this query. In addition the query is hardened to prevent missing data breaking the migration.

## Why it's being changed

These measures are intended to improve performance when running the `dotdigital:migrate` command, for merchants with large volumes of customers and subscribers. We already use `NOT EXISTS` in the   `InsertEmailContactTableCustomerSales` migration.

## How to review / test this change

- Before switching to this branch, run `bin/magento dotdigital:migrate`
- Take a note of the numbers of customers and subscribers you migrated
- Run `bin/magento setup:upgrade`
- Re-run `bin/magento dotdigital:migrate`
- Confirm the numbers match

## Notes

We experimented with adding a composite index on the `email` and `store_id` columns in the `email_contact` table, but in our performance testing found this to have minimal impact.

I should also note that the actual query optimisation here is minor. The merchant involved had ~50k rows in `newsletter_subscriber` with no email, which led to an infinite loop during the migration. The changes in this PR protect against this happening, so in the end they are more valuable than any performance boost from the NOT EXISTS (which is negligible).

Related work items: #308837