Merged PR 68044: Allow contact id for row matching any eligible store id
## What's being changed

The join condition restricting order sync to contacts with contact ids in the correct scope.

## Why it's being changed

Contacts in email_contact will take the store_id of the store they registered on. But, contact ids are essentially website scoped because we connect DD at website level. So we want to restrict order sync to rows that have a contact id, for _any_ of the store ids belonging to the website we are syncing orders for.

The original condition added was too restrictive: https://github.com/dotmailer/dotmailer-magento2-extension/commit/2b22edfd73cf073925439c208f9baeff62219424.

## How to review / test this change

- Create a website with 2 stores
- Create a customer on store 1
- Allow the customer to sync and have a contact id stored in the table
- Place an order on store 1
- Sync the order
- Place an order on store 2
- Sync the order
- Both orders should be synced

Related work items: #312618